Use flyway migrations to deploy schema changes
This change also renames the table and column names to match our naming
conventions (which also match the Scala conventions).

There is documentation for naming database objects, and an updated .gitignore
for the Idea '.idea' project file, vim .*.swp and patch *.orig files

This adds the command "sbt core/flywayMigrate" and friends.

The 'init' method for schema creation is removed, because schema creation is
done using flyway migrations now.