Load/save packages
Create a 'Package' class that maps to the 'Package' SQL table. This supports
listing the packages with GET /packages (commented out) and adding a package with POST
/packages.

This commit introduces the following whitelisted interactions:
WL-8 External Resolver persists new package data to External Resolver Database
WL-5 Web Server sends new package data to External Resolver

There isn't a whitelisted interaction for returning the list of packages, so
that functionality is disabled in this commit.

BUGS

Returns a 500 rather than 409 HTTP status code when attempting to create a
duplicate package.