Make proxy behavior more strict
Previously the proxy required that both the core and resolver gave a 'success'
response before it would return a success to the user.

This change changes the rules to require that at least one upstream server
returns a 404 Not Found response. The other is returned to the user.

Vins must be duplicated between core and resolver, and so these are changed to
return 204 'No Content' responses. The proxy will accept a pair of 204
responses as valid.  There is the possibility here that the two will get our of
sync, but some changes coming from Txus will fix this by making Vin creation
idempotent.

The logic for this is in RequestHelpers (and is tested in RequestHelpersSpec).

This change also factors out common code that turns a WSResponse into a Play Result.