Log all requests to the backend
This implements mitigation C04 "Log transactions to and from SOTA Server".  The
current user is a placeholder at the moment, because user logins have not been
implemented yet.

The log messages are written out to an 'audit' log, so they can easily be
stripped out of the main log stream and kept somewhere safe.