Support different data types communicating with client
When parsing/serializing request/responses, support a mapping from
server representations to client representations and vice versa.

This fixes a bug when creating campaigns, as the web client is not able
to send a namespace, which should be infered by the server.

In principle whe client representation should not be tied to the server
representation or a db schema, so this PR goes in that direction.