Add new endpoint to change package install pos
This allows the api client to set the install order of a particular set
of updates to a specific vehicle.

The new endpoint is available at `api/v1/vehicle_updates/vin/order`,
this accepts a `PUT` request with the format:

     {
        "0": "UUID for update request",
        "1": "UUID for update request"
        // ...
     }

The packages will be delivered in this order when the client queries the
`api/v1/vehicle_updates` API. The install pos will be returned in a new
attribute that should be used by the client to sort updates to install.

    GET api/v1/vehicle_updates/

    [{
      "createdAt": "2016-04-27T02:00:00.000+02:00",
      "installPos": 0,
      "packageId": {
        "name": "pkgname",
        "version": "1.0.0"
      },
      "requestId": "41078e2f-b8c5-4fda-87fa-1f4761413553"
    }]