Do not store upload file in memory before upload
For local storage, the file is copied directly to disk without being
staged in memory.

For s3 storage, amazon requires us to specify the size of the file being
uploaded, so we need to first store the file in disk before uploading,
so we can calculate it's size. The file is now stored in disk for the
upload, and deleted immediately after the upload is finished.

Increased max file size and idle time to support large uploads

Refactored Package resource