Use a simple pk for Package
Maintains uniqueness on previous composite pk, (namespace, name,
version)

This simplifies all queries and code related to packages as namespace,
name and version are now simple attributes of a Package entity, so the
code does not need to deal with them if they will not be used.

This simplification also made it possible to remove the `namespace`
column on all tables except Package, all other entities are linked by a
Package, so this is redundant.